Mejoras 14102025

+++ SYSRENT.Domain/Horario/DTO/DtoConsulCatHorario.cs
@@ -2,7 +2,7 @@ namespace SYSRENT.Domain.Horario.DTO;
 
 public sealed record DtoConsulCatHorario
 (
-    int IdHorario,
-    string Descrip,
-    int Minutos
+  byte IdHorario,
+  string Descrip,
+  byte Minutos
 );
